Rename metron_tagger.py to main.py
